# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Badminton Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Player {
  id: number;
  name: string;
  avatarColor: string;
  createdAt: string;
}

export interface CreatePlayerRequest {
  name: string;
  avatarColor: string;
}

export type MatchMatchType =
  (typeof MatchMatchType)[keyof typeof MatchMatchType];

export const MatchMatchType = {
  singles: "singles",
  doubles: "doubles",
} as const;

export interface MatchPlayer {
  playerId: number;
  playerName: string;
  avatarColor: string;
  /** 1 or 2 */
  team: number;
}

export interface Match {
  id: number;
  matchType: MatchMatchType;
  team1Score: number;
  team2Score: number;
  /** 1 or 2 */
  winnerTeam: number;
  players: MatchPlayer[];
  notes?: string;
  playedAt: string;
  createdAt: string;
}

export interface PlayerStats {
  playerId: number;
  playerName: string;
  totalMatches: number;
  wins: number;
  losses: number;
  winRate: number;
  singlesWins: number;
  singlesLosses: number;
  doublesWins: number;
  doublesLosses: number;
  currentWinStreak: number;
  longestWinStreak: number;
  recentMatches: Match[];
}

export interface LeaderboardEntry {
  playerId: number;
  playerName: string;
  avatarColor: string;
  totalMatches: number;
  wins: number;
  losses: number;
  winRate: number;
  currentWinStreak: number;
}

export type CreateMatchRequestMatchType =
  (typeof CreateMatchRequestMatchType)[keyof typeof CreateMatchRequestMatchType];

export const CreateMatchRequestMatchType = {
  singles: "singles",
  doubles: "doubles",
} as const;

export interface CreateMatchRequest {
  matchType: CreateMatchRequestMatchType;
  team1PlayerIds: number[];
  team2PlayerIds: number[];
  team1Score: number;
  team2Score: number;
  notes?: string;
  playedAt: string;
}

export type GetMatchesParams = {
  limit?: number;
  offset?: number;
};