# lib/api-zod/src/generated/types/createMatchRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Badminton Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateMatchRequestMatchType } from "./createMatchRequestMatchType";

export interface CreateMatchRequest {
  matchType: CreateMatchRequestMatchType;
  team1PlayerIds: number[];
  team2PlayerIds: number[];
  team1Score: number;
  team2Score: number;
  notes?: string;
  playedAt: Date;
}