# lib/api-zod/src/generated/types/createPlayerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Badminton Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePlayerRequest {
  name: string;
  avatarColor: string;
}