# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Badminton Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  playerId: number;
  playerName: string;
  avatarColor: string;
  totalMatches: number;
  wins: number;
  losses: number;
  winRate: number;
  currentWinStreak: number;
}