# lib/api-zod/src/generated/types/match.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Badminton Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchMatchType } from "./matchMatchType";
import type { MatchPlayer } from "./matchPlayer";

export interface Match {
  id: number;
  matchType: MatchMatchType;
  team1Score: number;
  team2Score: number;
  /** 1 or 2 */
  winnerTeam: number;
  players: MatchPlayer[];
  notes?: string;
  playedAt: Date;
  createdAt: Date;
}