# lib/api-zod/src/generated/types/matchPlayer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Badminton Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface MatchPlayer {
  playerId: number;
  playerName: string;
  avatarColor: string;
  /** 1 or 2 */
  team: number;
}