# lib/api-zod/src/generated/types/playerStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Badminton Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { Match } from "./match";

export interface PlayerStats {
  playerId: number;
  playerName: string;
  totalMatches: number;
  wins: number;
  losses: number;
  winRate: number;
  singlesWins: number;
  singlesLosses: number;
  doublesWins: number;
  doublesLosses: number;
  currentWinStreak: number;
  longestWinStreak: number;
  recentMatches: Match[];
}